Create an abstract class for pool implementation
To ease the integration of futur pool implementation.
