+++ .github/workflows/ci.yml
@@ -35,7 +35,7 @@ jobs:
 
       - name: Test & Coverage
         if: github.ref == 'refs/heads/master'
-        run: | 
+        run: |
           npm run test
           export COVERALLS_REPO_TOKEN=${{ secrets.COVERALLS_REPO_TOKEN }}
           npm run coverage
