+++ src/index.ts
@@ -2,11 +2,8 @@ import DynamicThreadPool from './dynamic'
 import FixedThreadPool from './fixed'
 import { ThreadWorker } from './workers'
 
-export {
-  Draft,
-  FixedThreadPoolOptions,
-  WorkerWithMessageChannel
-} from './fixed'
+export { Draft, WorkerWithMessageChannel } from './pool'
+export { FixedThreadPoolOptions } from './fixed'
 export { DynamicThreadPoolOptions } from './dynamic'
 export { ThreadWorkerOptions } from './workers'
 export { FixedThreadPool, DynamicThreadPool, ThreadWorker }
